partner_bot/requirements.txt
aiogram==2.21
python-dotenv~=0.20.0